Update for Python2.7

+++ python/src/nnabla/utils/converter/nnablart/nnb.py
@@ -26,7 +26,7 @@ from .utils import create_nnabart_info
 
 class NnbExporter:
     def _align(self, size):
-        return math.ceil(size / 4) * 4
+        return int(math.ceil(size / 4) * 4)
 
     def _alloc(self, size=-1, data=b''):
         size = len(data) if size < 0 else size
